+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 ## Breaking changes
 
+-   Changed keyword argument `u` for inputs (when evaluating an object) to `params` 
 -   Removed "set external temperature" and "set external potential" options. Use "external submodels" option instead ([#862](https://github.com/pybamm-team/PyBaMM/pull/862))
 
 # [v0.2.0](https://github.com/pybamm-team/PyBaMM/tree/v0.2.0) - 2020-02-26